API talking and partially writing
The API is giving us the right information and it is writing in the correct places. However, KeyError is an issue.
I believe this is when it is a direct match and there is no dictionary within the dictionary. Should be possible to solve using .get() to set a default key.